move location tests outside of general consolidate tests; be more specific as to what args are allowed, fix docstrings
